add workflows add check owner available slot for each service

+++ app/controllers/api/v1/spa_n_salons_controller.rb
@@ -1,10 +1,7 @@
-# frozen_string_literal: true
-
 module Api
   module V1
     class SpaNSalonsController < ApplicationController
       before_action :auth
-      before_action :owner_iscorrect, only: %i[create update destroy]
       before_action :set_spa_n_salon, only: %i[show update destroy]
 
       # GET /spa_n_salons
@@ -19,52 +16,56 @@ module Api
 
       # GET /spa_n_salons/1
       def show
-        @spa_n_salon
         @services = @spa_n_salon.services.order('cost asc')
-        
+
         render json: @services
       end
 
       # POST /spa_n_salons
       def create
-        @spa_n_salon = SpaNSalon.new(spa_n_salon_params)
-
-        if @spa_n_salon.save
-          render json: @spa_n_salon, status: :ok
+        if owner_is_correct?
+          @spa_n_salon = SpaNSalon.new(spa_n_salon_params)
+          if @spa_n_salon.save
+            render json: @spa_n_salon, status: :ok
+          else
+            render json: @spa_n_salon.errors, status: :unprocessable_entity
+          end
         else
-          render json: @spa_n_salon.errors, status: :unprocessable_entity
+          render json: { message: "unauthorized owner "}, status: :ok
         end
       end
 
       # PATCH/PUT /spa_n_salons/1
       def update
-        if @spa_n_salon.update(spa_n_salon_params)
-          render json: @spa_n_salon
+        if owner_is_correct?
+          if @spa_n_salon.update(spa_n_salon_params)
+            render json: @spa_n_salon
+          else
+            render json: @spa_n_salon.errors, status: :unprocessable_entity
+          end
         else
-          render json: @spa_n_salon.errors, status: :unprocessable_entity
+          render json: { message: "unauthorized owner "}, status: :ok
         end
+        
       end
 
       # DELETE /spa_n_salons/1
       def destroy
-        if @spa_n_salon.destroy
-          render json: { message: "deleted successfully" }, status: :ok
+        if owner_is_correct?
+          if @spa_n_salon.destroy
+            render json: { message: 'deleted successfully' }, status: :ok
+          else
+            render json: @spa_n_salon.errors, status: :unprocessable_entity
+          end 
         else
-          render json: @spa_n_salon.errors, status: :unprocessable_entity
+          render json: { message: "unauthorized owner "}, status: :ok
         end
       end
-      
+
       def sorted_service
-        @salon = SpaNSalon.joins(:services)
-        @service = []
-        @salon.each do |salon|
-          @service << salon.services.order('cost asc')
-        end
-        render json: @se
-      end
-      
-      def working_dayTime
-        SpaNSalon.join(:work_schedules)
+        @salon = SpaNSalon.joins('INNER JOIN Services on spa_n_salons.id = services.spa_n_salon_id').select('*').order('cost asc')
+
+        render json: @salon
       end
 
       private
@@ -74,11 +75,14 @@ module Api
         @spa_n_salon = SpaNSalon.find(params[:id])
       end
 
-      def owner_iscorrect
-        spa_id = params[:id]
-        @owner = SpaNSalons.where(id: spa_id).pluck(:owner_id)
-        @current_owner = current_user.owner if current_user
-        @owner == @current_owner
+      def owner_is_correct?
+        spa_id = params[:spa_n_salon_id]
+        @owner = SpaNSalon.where(id: spa_id).pluck(:owner_id)
+        if current_user
+          @current_owner = Owner.where(user_id: current_user.id) 
+          return true if (@owner == @current_owner)  
+        end
+        false
       end
 
       # Only allow a list of trusted parameters through.
