+++ app/controllers/api/v1/work_schedules_controller.rb
@@ -1,10 +1,7 @@
-# frozen_string_literal: true
-
 module Api
   module V1
     class WorkSchedulesController < ApplicationController
       before_action :auth
-      before_action :owner_is_correct, only: %i[create update destroy]
       before_action :set_work_schedule, only: %i[show update destroy]
 
       # GET /work_schedules
@@ -21,29 +18,42 @@ module Api
 
       # POST /work_schedules
       def create
-        @work_schedule = WorkSchedule.new(work_schedule_params)
-        if @work_schedule.save
-          render json: @work_schedule, status: :ok
+        if owner_is_correct?
+          @work_schedule = WorkSchedule.new(work_schedule_params)
+          if @work_schedule.save
+            render json: @work_schedule, status: :ok
+          else
+            render json: @work_schedule.errors, status: :unprocessable_entity
+          end
         else
-          render json: @work_schedule.errors, status: :unprocessable_entity
+          render json: { message: "unauthorized owner "}, status: :ok
         end
+        
       end
 
       # PATCH/PUT /work_schedules/1
       def update
-        if @work_schedule.update(work_schedule_params)
-          render json: @work_schedule
+        if owner_is_correct?
+          if @work_schedule.update(work_schedule_params)
+            render json: @work_schedule
+          else
+            render json: @work_schedule.errors, status: :unprocessable_entity
+          end
         else
-          render json: @work_schedule.errors, status: :unprocessable_entity
+          render json: { message: "unauthorized owner "}, status: :ok
         end
       end
 
       # DELETE /work_schedules/1
       def destroy
-        if @work_schedule.destroy
-          render json: { message: 'work schedule deleted successfully' }, status: :ok
+        if owner_is_correct?
+          if @work_schedule.destroy
+            render json: { message: 'work schedule deleted successfully' }, status: :ok
+          else
+            render json: @work_schedule.errors, status: :unprocessable_entity
+          end
         else
-          render json: @work_schedule.errors, status: :unprocessable_entity
+          render json: { message: "unauthorized owner "}, status: :ok
         end
       end
 
@@ -52,6 +62,16 @@ module Api
       # Use callbacks to share common setup or constraints between actions.
       def set_work_schedule
         @work_schedule = WorkSchedule.find(params[:id])
+      end
+      
+      def owner_is_correct?
+        spa_id = params[:spa_n_salon_id]
+        @owner = SpaNSalon.where(id: spa_id).pluck(:owner_id)
+        if current_user
+          @current_owner = Owner.where(user_id: current_user.id) 
+          return true if (@owner == @current_owner)  
+        end
+        false
       end
 
       # Only allow a list of trusted parameters through.
